requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
qdrant-client==1.7.0
sentence-transformers==2.3.1
pypdf2==3.0.1
python-docx==1.1.0
pandas==2.1.4
reportlab==4.0.7
beautifulsoup4==4.12.2
lxml==5.1.0
pydantic==2.5.3
aiohttp==3.9.1